Closes #29: Disallow copy / assignment constructor
The `M5EPD_Canvas` class defines an assignment constructor (wihout
defining a copy constructor). The original goal was to avoid
ownership issues because of dangling sprites. However, from a user
perspective it is not clear that this behvior is not desired.

To force the user not to use this way of constructing a canvas, this
patch deletes the copy and assignment constructor.